fix and refactoring
- fix issue #5
- remove Logging.(cpp|h)
