--- Plugins/ModuleGeneration/Resources/{ModuleName}/Private/Logging.cpp
@@ -1,5 +1,0 @@
-// {Copyright}
-
-#include "Logging.h"
-
-DEFINE_LOG_CATEGORY({ModuleName});
