--- Plugins/ModuleGeneration/Resources/{ModuleName}/Private/Logging.h
@@ -1,7 +1,0 @@
-// {Copyright}
-
-#pragma once
-
-#include "CoreMinimal.h"
-
-DECLARE_LOG_CATEGORY_EXTERN({ModuleName}, Log, All);
