--- Plugins/ModuleGeneration/Resources/{ModuleName}/Public/{ModuleName}.h
@@ -1,16 +1,0 @@
-// {Copyright}
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Modules/ModuleManager.h"
-
-class F{ModuleName} : public IModuleInterface
-{
-public:
-
-	/** IModuleInterface implementation */
-	void StartupModule() override;
-	void ShutdownModule() override;
-	
-};
