+++ Plugins/ModuleGeneration/Source/ModuleGeneration/Public/ModuleGeneration.h
@@ -10,8 +10,8 @@ class FModuleGenerationModule : public IModuleInterface
 public:
 
 	/** IModuleInterface implementation */
-	void StartupModule() override;
-	void ShutdownModule() override;
+	virtual void StartupModule() override;
+	virtual void ShutdownModule() override;
 
 private:
